requirements.txt
# Основные зависимости
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.4.2
scipy==1.10.1

# Обработка текста
razdel==0.5.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844

# Нейросети
torch==2.2.1
torchvision==0.17.1

# Визуализация
matplotlib==3.7.1
seaborn==0.12.2

# Утилиты
tqdm==4.65.0
pyyaml==6.0.1
jupyter==1.0.0
ipython==8.12.0

# Для загрузки данных
datasets==2.14.5
requests==2.31.0